# services/content/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: Content Service API
  version: 1.0.0
  description: |
    Production-grade content management service for the Suuupra platform.
    
    ## Features
    - Content CRUD operations with rich metadata
    - Large file uploads with S3 multipart and progress tracking
    - Elasticsearch-powered search and discovery
    - Content approval workflow with versioning
    - Real-time WebSocket notifications
    - Multi-tenant architecture with RBAC
    
    ## Authentication
    All endpoints require JWT authentication via the `Authorization: Bearer <token>` header.
    
    ## Rate Limiting
    API endpoints are rate limited. Check response headers for current limits.
    
    ## Versioning
    Content uses semantic versioning (major.minor.patch) with immutable history.
  contact:
    name: Suuupra Platform Team
    email: platform-team@suuupra.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8082/api/v1
    description: Development server
  - url: https://api.suuupra.com/content/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /content/{id}/submit-approval:
    post:
      summary: Submit for approval
      description: Submit content for approval workflow
      operationId: submitForApproval
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '200':
          description: Content submitted for approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/content/{id}/approve:
    post:
      summary: Approve content
      description: Approve content (moderator/admin only)
      operationId: approveContent
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowActionRequest'
      responses:
        '200':
          description: Content approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/content/{id}/reject:
    post:
      summary: Reject content
      description: Reject content with reason (moderator/admin only)
      operationId: rejectContent
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/WorkflowActionRequest'
                - type: object
                  required: [reason]
      responses:
        '200':
          description: Content rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/content/{id}/publish:
    post:
      summary: Publish content
      description: Publish approved content (moderator/admin only)
      operationId: publishContent
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/WorkflowActionRequest'
                - type: object
                  properties:
                    versionBump:
                      $ref: '#/components/schemas/VersionBump'
      responses:
        '200':
          description: Content published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ContentId:
      name: id
      in: path
      required: true
      description: Content ID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ContentType:
      type: string
      enum: [video, article, quiz, document]
    ContentStatus:
      type: string
      enum: [draft, pending_approval, approved, published, archived]
    VersionBump:
      type: string
      enum: [major, minor, patch]

    # Core Models
    Content:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
        title:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        contentType:
          $ref: '#/components/schemas/ContentType'
        status:
          $ref: '#/components/schemas/ContentStatus'
        version:
          type: string
          pattern: '^\\d+\\.\\d+\\.\\d+$'
        categoryId:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        metadata:
          type: object
          additionalProperties: true
        fileInfo:
          $ref: '#/components/schemas/FileInfo'
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        etag:
          type: string
      required:
        - id
        - tenantId
        - title
        - contentType
        - status
        - version
        - createdBy
        - createdAt
        - updatedAt
        - etag
    FileInfo:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
          minimum: 1
        s3Key:
          type: string
        cdnUrl:
          type: string
          format: uri
        checksumSha256:
          type: string
          pattern: '^[a-fA-F0-9]{64}$'
        uploadedAt:
          type: string
          format: date-time
      required:
        - filename
        - contentType
        - fileSize
        - s3Key
        - checksumSha256
        - uploadedAt

    # Request/Response Schemas
...
    WorkflowActionRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
        metadata:
          type: object
          additionalProperties: true

    # Response Schemas
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      required:
        - success
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - code
        - message
    ResponseMeta:
      type: object
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
      required:
        - requestId
        - timestamp
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
    ContentResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Content'
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
          required:
            - success
            - error
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                errors:
                  - field: title
                    message: must not be empty
                    value: ""
...
    Forbidden:
      description: Insufficient privileges
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Insufficient privileges
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Content not found